Fix the handling of pending_xref objects.
Previously, when a pending_xref was getting created, it was creating
with a reference node in it (with essentially the same target) and then
its children were contained with that. When sphinx would render this, it
would create 2 <a> tags, one with no children but with a correctly
resolved href and the other with the correct children but the href is
not resolved.
This makes it so that it creates pending_xref nodes the same way sphinx
does when parsing rst. It creates the pending_xref with the content
node(s) as children. During pending_xref resolution, it replaces that
pending_xref node with a reference node containing the same children.