Allow the destination of a link to be determined by its string content.
If the destination of a link is empty, use the value of the string
contained in the text to be rendered, if there is any.
This makes it possible to do things like:

```
This is a link to [asyncio.Event]().
```

instead of:
```
This is a link to [asyncio.Event](asyncio.Event).
```